Move dependencies to gemspec
Before this commit, the gem works properly when working with the
source code, but encounters some problems with dependencies when
installing via rubygems.

This will move all dependencies into the gemspec and refer to the
gemspec from the Gemfile in order for the gem to work similarly when
run from source or as a gem.