Refactor the initialize method to allow different file formats
The fast_gettext gem which this library wraps supports multiple formats
for translation files. This commit allows the user to specify which
format to use for loading translations. This is needed to give us
flexibility to choose whether to use .mo binary files (more overhead,
but parallels the process for other languages) or read directly from .po
files.