Allow restriction of extracted comments
When no argument is passed to rxgettext's `--add-comments` flag, all
comments are extracted. We want to be able to restrict this to just
comments made for translators, so we allow the user to supply a tag
using a new `comments_tag` config in the config.yaml. This tag will
default to "TRANSLATORS" if not supplied. If supplied but empty,
all comments will be extracted, affording us backwards-compatibility.

Fixes https://tickets.puppetlabs.com/browse/INTL-9