(PUP-7122) Add a rake task for diffing POT files
This commit adds a new rake task which will cache the existing POT file
of a project, generate a new one, and check for changes in actual string
content, only replacing the file if relevant changes are found. This
allows us to only update the POT file if necessary and avoid churn due
to changing modification dates.