(INTL-13) Make update_pot task also replace POT when strings are deleted
Previously the update_pot rake task only replaced the POT file when
strings were added or changed, but not when they were deleted. This
commit updates the comparison logic to account for the warnings issued
by msgcmp when strings have been deleted, and act on them to replace the
POT file.