(INTL-18) default_available_locales as global variable
Currently, available_locales is being overwritten every time a
module is initialized with GettextSetup, so if we want a German
translation, it is only be available if it is available in the
most recent module that has been initialized (alphabetical order).
This commit moves FastGettext.default_available locales to be
initialized outside of the GettextSetup initialization and
subsequently added to using a join every time a new project is
processed by GettextSetup. This allows us to have a 'master'
list of locales to choose from, which will contain all locales
from all projects, instead of relying on the last initialized
project to contain all the locales we might need. Spec tests
include testing that a list is built when more than one project
is initialized and that locales from that list can be set.