(INTL-17) Implement text_domain chaining
Instead of using individual text_domains, we will now be using a 'chain' which allows the continued use of _() instead of switching to d_() or D_(). A chain will aggregate all the text domains into one master text domain so that we will no longer need to switch domains to find a translation.