(INTL-25) Add check for string addition warning from msgcmp
This commit adds an additional regex to the update_pot rake task, which
checks for the warning issued by `msgcmp` when it sees strings in the
new POT file which were not present in the old one. Previously it was
assumed `msgcmp` exited 1 in this case, but it does not, so the
additional grepping of stderr is necessary to detect string additions.