(INTL-25) Add tests for pot file comparison
This commit adds tests to make sure we can properly detect all the
different kinds (and absence) of POT file string changes. In order to
facilitate this testing, the helper methods around the `gettext` rake
tasks have been moved into their own file. Note that these new tests
are skipped if msgcmp is not present on the system.