When negotiating locale, only consider language, but not country variant
When the Accept header contains something like 'de-DE', and we have 'en'
and 'de' available, we used to only look for a 'de-DE' message catalog, but
not for the more general 'de' catalog.

This patch makes it so that we drop country variant before looking for a
message catalog, i.e., we treat 'de-DE' and 'de' as exactly the same. This
will work as long as we do not need to have both a message catalog for the
language and one for that language respecting a variant.

Fixes https://tickets.puppetlabs.com/browse/ENTERPRISE-1060