(PUP-7881) Put default locales at the front of available_locales
FastGettext prefers the first item in its list of available locales when
choosing a locale before it has been set by the user. However, we were
putting our defaults at the end of the list we use to initialize
FastGettext, meaning FastGettext was choosing an essentially random
locale as its initial value for `locale`. This commit updates the way we
 construct our `available_locales` list to put the defaults first.