(FM-6415) create header_only option for pot generation
for the sake of consistency across POT files (ruby, puppet, metadata), it seems to make sense for the headers to all be created in one place. This change allows for the creation of 'blank' POT files, or POT files with only a header, by passing in a blank input file. This provides the following benefits:
- When POT files are merged, differences in header content can cause a lot of clutter in the final POT file. identical headers will cut down on this clutter substantially.
- The generate function can be used to create blank POT files to allow for the appending of POT files for other languages in the future
- Consistency is great
- GettextSetup knows a lot about the current project so it can provide header info that other libraries may not be able to, like project name, etc.