(INTL-31) Refactor rake tasks to not setup FastGettext runtime
Prior to this commit the Gettext:Pot module would setup the FastGettext
config and runtime. This would cause strange conflicts with other calls
to FastGettext (for example, when calling the rake task in an
environment that also contains Puppet 5.3.3) so we decided that
since the rake tasks do not require the FastGettext runtime config,
they should not try to set it up.
This commit refactors the GettextSetup.initialize method to call out
to a GettextSetup.initialize_config method, that only configures the
file-related info for the library. Then the GettextSetup:Pot module
is updated to call the GettextSetup.initialized_config method instead
of initialize. As far as I know, the GettextSetup:Pot module is only
called by the rake tasks so hopefully this is a safe fix.