(maint) Use new pot if msgcmp unavailable
If msgcmp isn't on the system, we were previously relying on the system
throwing an IOError that we could catch. Unfortunately, this wasn't
consistently happening. This commit adds an additional check to
explicitly print out a warning if msgcmp is not available and to
proceed with replacing the old pot with the new one.